refactor: move init validation out of flattened_dataset
