+++ src/application/t2d/check_flattened_ds_integrity.py
